adding in notebook for calibrating the heara golend data set.
This is not an initial commit and the notebook has gone through many iterations.

This iteration looked at just using the mean for the firstcal solutions for all of the solutions. This took away the time dependence of the solutions. Hoping to fix offsets of 43_89 by pi, but did not work.